Refactor::프로젝트 폴더 구조 리펙토링
