+++ src/components/WordSet/WordSet.css
