+++ src/pages/MemoSet/MemoSet.tsx
@@ -1,11 +1,11 @@
 import React, { LegacyRef, RefObject, useEffect, useState } from "react";
-import Header from "../components/Header";
-import WordSet from "../components/WordSet";
+import Header from "../../components/Header/Header";
+import WordSet from "../../components/WordSet/WordSet";
 import Popup from "reactjs-popup";
 import "../styles/MemoSet.css";
 import { IoAddOutline } from "react-icons/io5";
 import { Link, useHref, useLocation } from "react-router-dom";
-import { useHorizontalScroll } from "../hooks/useSideScroll";
+import { useHorizontalScroll } from "../../hooks/useSideScroll";
 import axios from "axios";
 
 interface IProp {
