+++ src/pages/Memorize/Memorize.css
