+++ src/pages/Search/Search.tsx
@@ -2,7 +2,7 @@ import axios, { AxiosResponse } from "axios";
 import React, { useEffect, useState } from "react";
 import { set } from "react-hook-form";
 import { Link, useLocation } from "react-router-dom";
-import Header from "../components/Header";
+import Header from "../../components/Header/Header";
 import "../styles/Search.css";
 
 interface IProp {
