+++ src/pages/SharedSetWord.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
-import Header from "../components/Header";
+import Header from "../components/Header/Header";
 import "../styles/wordList.css";
 import SharedWord from "../components/SharedWord";
 import axios from "axios";
